# swagger.json
{"paths": {"/taxis": {"get": {"description": "Get taxi list", "parameters": [{"in": "query", "name": "page", "schema": {"type": "integer"}, "required": false, "description": "Page number of results"}, {"in": "query", "name": "per_page", "schema": {"type": "integer"}, "required": false, "description": "Number of results per page"}], "responses": {"200": {"description": "Returns the list of taxis", "content": {"application/json": {"schema": {"type": "object", "properties": {"taxis": {"type": "array", "items": {"$ref": "#/components/schemas/TaxiModel"}}, "count": {"type": "integer"}}}}}}}}}, "/trajectories": {"get": {"description": "Get trajectories list", "parameters": [{"in": "query", "name": "taxiId", "schema": {"type": "integer"}, "required": true, "description": "taxi id"}, {"in": "query", "name": "date", "schema": {"type": "string"}, "required": true, "description": "Date format DD-MM-YYYY"}, {"in": "query", "name": "page", "schema": {"type": "integer"}, "required": false, "description": "Page number of results"}, {"in": "query", "name": "per_page", "schema": {"type": "integer"}, "required": false, "description": "Number of results per page"}], "responses": {"200": {"description": "Return trajectories list for a specific taxi", "content": {"application/json": {"schema": {"type": "object", "properties": {"trajectories": {"type": "array", "items": {"$ref": "#/components/schemas/trajectoriesModel"}}}}}}}}}}, "/trajectories/latest": {"get": {"description": "Get last location (trajectory) of all taxis", "parameters": [{"in": "query", "name": "page", "schema": {"type": "integer"}, "required": false, "description": "Page number of results"}, {"in": "query", "name": "per_page", "schema": {"type": "integer"}, "required": false, "description": "Number of results per page"}], "responses": {"200": {"description": "Returns the list of last location of all taxis", "content": {"application/json": {"schema": {"type": "object", "properties": {"taxis": {"type": "array", "items": {"$ref": "#/components/schemas/trajectoriesModel"}}}}}}}}}}}, "info": {"title": "Fleet Management API", "version": "1.0.0"}, "openapi": "3.0.2", "components": {"schemas": {"TaxiModel": {"properties": {"id": {"type": "integer", "format": "int64"}, "plate": {"type": "string"}}}, "LocationsModel": {"properties": {"id": {"type": "integer", "format": "int64"}, "plate": {"type": "string"}, "timestamp": {"type": "string", "format": "date-time"}, "lat": {"type": "string"}, "lon": {"type": "string"}}}}}}